requirements.txt
django
djangorestframework
django-cors-headers
requests
django-channels
pillow
daphne
urllib3
google-auth
django-allauth
gunicorn
fuzzywuzzy
celery
redis
django-celery
python-dotenv
psycopg2
channels[daphne]